% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Queries historical exchange rates for any 'base' and 'symbol' currency pair. When no params passed check EUR - HUF}
\usage{
get_exchange_rates(
  base_arg = "EUR",
  symbol_arg = "HUF",
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{base_arg}{base currency argument}

\item{symbol_arg}{to which symbol argument}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function has been retried}
}
\value{
\code{data.table} object
}
\description{
Queries historical exchange rates for any 'base' and 'symbol' currency pair. When no params passed check EUR - HUF
}
\examples{
get_exchange_rates()
get_exchange_rates('EUR', 'USD', start_date = '2020-05-12', end_date = '2020-05-13')
}
