% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R, R/get_usdhuf.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Look up the value of a US dollar in Hungarian forints}
\usage{
get_usdhuf(retried = 0)

get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
number

number
}
\description{
Look up the value of a US dollar in Hungarian forints

Look up the value of a US dollar in Hungarian forints
}
\examples{
get_usd(42)
get_usd(42)
}
