% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R, R/get_usdhufs.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Look up the historical value of a US Dollar in Hungarian Forints}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)

get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object

\code{data.table} object
}
\description{
Look up the historical value of a US Dollar in Hungarian Forints

Look up the historical value of a US Dollar in Hungarian Forints
}
\examples{
get_usdhufs()
get_usdhufs(start_date = "2021-04-18", end_date = "2021-05-18")
}
